refactor(progress): standardize reporter protocol imports
