+++ src/frame_compare/orchestration/context.py
@@ -17,7 +17,7 @@ from frame_compare.vs.types import HDRMetadata
 if TYPE_CHECKING:
     from frame_compare.analysis.types import SelectionBreakdown
     from frame_compare.config.schema import ConfigSchema
-    from frame_compare.utils.progress import ProgressReporter
+    from frame_compare.utils.progress_protocol import ProgressReporter
     from frame_compare.utils.types import WorkspacePaths
